+++ bindings/nostr-js/src/event/mod.rs
@@ -70,8 +70,8 @@ impl JsEvent {
     }
 
     #[wasm_bindgen(getter)]
-    pub fn kind(&self) -> f64 {
-        self.inner.kind().as_f64()
+    pub fn kind(&self) -> u16 {
+        self.inner.kind().as_u16()
     }
 
     #[wasm_bindgen(getter)]