+++ crates/nostr/src/event/kind.rs
@@ -15,17 +15,17 @@ use serde::de::{Deserialize, Deserializer, Error, Visitor};
 use serde::ser::{Serialize, Serializer};
 
 /// NIP90 - Job request range
-pub const NIP90_JOB_REQUEST_RANGE: Range<u64> = 5_000..5_999;
+pub const NIP90_JOB_REQUEST_RANGE: Range<u16> = 5_000..5_999;
 /// NIP90 - Job result range
-pub const NIP90_JOB_RESULT_RANGE: Range<u64> = 6_000..6_999;
+pub const NIP90_JOB_RESULT_RANGE: Range<u16> = 6_000..6_999;
 /// Regular range
-pub const REGULAR_RANGE: Range<u64> = 1_000..10_000;
+pub const REGULAR_RANGE: Range<u16> = 1_000..10_000;
 /// Replaceable range
-pub const REPLACEABLE_RANGE: Range<u64> = 10_000..20_000;
+pub const REPLACEABLE_RANGE: Range<u16> = 10_000..20_000;
 /// Ephemeral range
-pub const EPHEMERAL_RANGE: Range<u64> = 20_000..30_000;
+pub const EPHEMERAL_RANGE: Range<u16> = 20_000..30_000;
 /// Parameterized replaceable range
-pub const PARAMETERIZED_REPLACEABLE_RANGE: Range<u64> = 30_000..40_000;
+pub const PARAMETERIZED_REPLACEABLE_RANGE: Range<u16> = 30_000..40_000;
 
 macro_rules! kind_variants {
     ($($name:ident => $value:expr, $doc:expr),* $(,)?) => {
@@ -49,38 +49,38 @@ macro_rules! kind_variants {
             /// Represents a parameterized replaceable event.
             ParameterizedReplaceable(u16),
             /// Represents a custom event.
-            Custom(u64),
+            Custom(u16),
         }
 
-        impl From<u64> for Kind {
-            fn from(u: u64) -> Self {
+        impl From<u16> for Kind {
+            fn from(u: u16) -> Self {
                 match u {
                     $(
                         $value => Self::$name,
                     )*
-                    x if (NIP90_JOB_REQUEST_RANGE).contains(&x) => Self::JobRequest(x as u16),
-                    x if (NIP90_JOB_RESULT_RANGE).contains(&x) => Self::JobResult(x as u16),
-                    x if (REGULAR_RANGE).contains(&x) => Self::Regular(x as u16),
-                    x if (REPLACEABLE_RANGE).contains(&x) => Self::Replaceable(x as u16),
-                    x if (EPHEMERAL_RANGE).contains(&x) => Self::Ephemeral(x as u16),
-                    x if (PARAMETERIZED_REPLACEABLE_RANGE).contains(&x) => Self::ParameterizedReplaceable(x as u16),
+                    x if (NIP90_JOB_REQUEST_RANGE).contains(&x) => Self::JobRequest(x),
+                    x if (NIP90_JOB_RESULT_RANGE).contains(&x) => Self::JobResult(x),
+                    x if (REGULAR_RANGE).contains(&x) => Self::Regular(x),
+                    x if (REPLACEABLE_RANGE).contains(&x) => Self::Replaceable(x),
+                    x if (EPHEMERAL_RANGE).contains(&x) => Self::Ephemeral(x),
+                    x if (PARAMETERIZED_REPLACEABLE_RANGE).contains(&x) => Self::ParameterizedReplaceable(x),
                     x => Self::Custom(x),
                 }
             }
         }
 
-        impl From<Kind> for u64 {
-            fn from(e: Kind) -> u64 {
+        impl From<Kind> for u16 {
+            fn from(e: Kind) -> u16 {
                 match e {
                     $(
                         Kind::$name => $value,
                     )*
-                    Kind::JobRequest(u) => u as u64,
-                    Kind::JobResult(u) => u as u64,
-                    Kind::Regular(u) => u as u64,
-                    Kind::Replaceable(u) => u as u64,
-                    Kind::Ephemeral(u) => u as u64,
-                    Kind::ParameterizedReplaceable(u) => u as u64,
+                    Kind::JobRequest(u) => u,
+                    Kind::JobResult(u) => u,
+                    Kind::Regular(u) => u,
+                    Kind::Replaceable(u) => u,
+                    Kind::Ephemeral(u) => u,
+                    Kind::ParameterizedReplaceable(u) => u,
                     Kind::Custom(u) => u,
                 }
             }
@@ -185,22 +185,22 @@ impl Hash for Kind {
 }
 
 impl Kind {
+    /// Get [`Kind`] as `u64`
+    #[inline]
+    pub fn as_u16(&self) -> u16 {
+        (*self).into()
+    }
+
     /// Get [`Kind`] as `u32`
     #[inline]
     pub fn as_u32(&self) -> u32 {
-        self.as_u64() as u32
+        self.as_u16() as u32
     }
 
     /// Get [`Kind`] as `u64`
     #[inline]
     pub fn as_u64(&self) -> u64 {
-        (*self).into()
-    }
-
-    /// Get [`Kind`] as `f64`
-    #[inline]
-    pub fn as_f64(&self) -> f64 {
-        self.as_u64() as f64
+        self.as_u16() as u64
     }
 
     /// Check if [`Kind`] is a NIP90 job request
@@ -208,7 +208,7 @@ impl Kind {
     /// <https://github.com/nostr-protocol/nips/blob/master/90.md>
     #[inline]
     pub fn is_job_request(&self) -> bool {
-        NIP90_JOB_REQUEST_RANGE.contains(&self.as_u64())
+        NIP90_JOB_REQUEST_RANGE.contains(&self.as_u16())
     }
 
     /// Check if [`Kind`] is a NIP90 job result
@@ -216,13 +216,13 @@ impl Kind {
     /// <https://github.com/nostr-protocol/nips/blob/master/90.md>
     #[inline]
     pub fn is_job_result(&self) -> bool {
-        NIP90_JOB_RESULT_RANGE.contains(&self.as_u64())
+        NIP90_JOB_RESULT_RANGE.contains(&self.as_u16())
     }
 
     /// Check if [`Kind`] is `Regular`
     #[inline]
     pub fn is_regular(&self) -> bool {
-        REGULAR_RANGE.contains(&self.as_u64())
+        REGULAR_RANGE.contains(&self.as_u16())
     }
 
     /// Check if [`Kind`] is `Replaceable`
@@ -231,31 +231,25 @@ impl Kind {
         matches!(self, Kind::Metadata)
             || matches!(self, Kind::ContactList)
             || matches!(self, Kind::ChannelMetadata)
-            || REPLACEABLE_RANGE.contains(&self.as_u64())
+            || REPLACEABLE_RANGE.contains(&self.as_u16())
     }
 
     /// Check if [`Kind`] is `Ephemeral`
     #[inline]
     pub fn is_ephemeral(&self) -> bool {
-        EPHEMERAL_RANGE.contains(&self.as_u64())
+        EPHEMERAL_RANGE.contains(&self.as_u16())
     }
 
     /// Check if [`Kind`] is `Parameterized replaceable`
     #[inline]
     pub fn is_parameterized_replaceable(&self) -> bool {
-        PARAMETERIZED_REPLACEABLE_RANGE.contains(&self.as_u64())
+        PARAMETERIZED_REPLACEABLE_RANGE.contains(&self.as_u16())
     }
 }
 
 impl fmt::Display for Kind {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.as_u64())
-    }
-}
-
-impl From<f64> for Kind {
-    fn from(kind: f64) -> Self {
-        Self::from(kind as u64)
+        write!(f, "{}", self.as_u16())
     }
 }
 
@@ -263,16 +257,16 @@ impl FromStr for Kind {
     type Err = ParseIntError;
 
     fn from_str(kind: &str) -> Result<Self, Self::Err> {
-        let kind: u64 = kind.parse()?;
+        let kind: u16 = kind.parse()?;
         Ok(Self::from(kind))
     }
 }
 
-impl Add<u64> for Kind {
+impl Add<u16> for Kind {
     type Output = Self;
 
-    fn add(self, rhs: u64) -> Self::Output {
-        let kind = self.as_u64();
+    fn add(self, rhs: u16) -> Self::Output {
+        let kind: u16 = self.as_u16();
         Kind::from(kind + rhs)
     }
 }
@@ -282,7 +276,7 @@ impl Serialize for Kind {
     where
         S: Serializer,
     {
-        serializer.serialize_u64(From::from(*self))
+        serializer.serialize_u16(self.as_u16())
     }
 }
 
@@ -291,7 +285,7 @@ impl<'de> Deserialize<'de> for Kind {
     where
         D: Deserializer<'de>,
     {
-        deserializer.deserialize_u64(KindVisitor)
+        deserializer.deserialize_u16(KindVisitor)
     }
 }
 
@@ -301,14 +295,21 @@ impl Visitor<'_> for KindVisitor {
     type Value = Kind;
 
     fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "an unsigned number")
+        write!(f, "a 16-bit unsigned number (0-65535)")
     }
 
-    fn visit_u64<E>(self, v: u64) -> Result<Kind, E>
+    fn visit_u16<E>(self, v: u16) -> Result<Self::Value, E>
     where
         E: Error,
     {
-        Ok(From::<u64>::from(v))
+        Ok(Self::Value::from(v))
+    }
+
+    fn visit_u64<E>(self, v: u64) -> Result<Self::Value, E>
+    where
+        E: Error,
+    {
+        Ok(Self::Value::from(v as u16))
     }
 }